pyproject.toml
[tool.poetry]
name = "repl-nix-rag-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.45.0"
flask = "^3.0.3"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.8.0.post1"
sentence-transformers = "^3.1.0"
markdown = "^3.7"
flake8 = "^7.1.1"
pyright = "^1.1.380"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
